refactor(payments): split payment schemas by domain
